+++ Exodus/src/utopia/exodus/database/access/single/DbDevice.scala
@@ -1,7 +1,6 @@
 package utopia.exodus.database.access.single
 
 import java.time.Instant
-import java.util.UUID.randomUUID
 
 import utopia.exodus.database.access.many.DbDescriptions
 import utopia.exodus.database.factory.device.DeviceKeyFactory
@@ -10,6 +9,7 @@ import utopia.exodus.database.model.user.UserDeviceModel
 import utopia.exodus.model.enumeration.StandardDescriptionRoleId
 import utopia.exodus.model.partial.DeviceKeyData
 import utopia.exodus.model.stored.DeviceKey
+import utopia.exodus.util.UuidGenerator
 import utopia.flow.generic.ValueConversions._
 import utopia.vault.database.Connection
 import utopia.vault.model.enumeration.ComparisonOperator.LargerOrEqual
@@ -156,8 +156,8 @@ object DbDevice
 			  * @param connection DB Connection (implicit)
 			  * @return This device key, now belonging to the specified user
 			  */
-			def assignToUserWithId(userId: Int)(implicit connection: Connection) = update(userId,
-				randomUUID().toString)
+			def assignToUserWithId(userId: Int)(implicit connection: Connection, uuidGenerator: UuidGenerator) =
+				update(userId, uuidGenerator.next())
 			
 			/**
 			  * Releases this device authentication key from the specified user, if that user is currently holding this key