+++ Exodus/src/utopia/exodus/rest/resource/device/DeviceKeyNode.scala
@@ -4,6 +4,7 @@ import utopia.access.http.Method.{Delete, Get}
 import utopia.access.http.Status.NotFound
 import utopia.exodus.database.access.single.{DbDevice, DbUser}
 import utopia.exodus.rest.util.AuthorizedContext
+import utopia.exodus.util.ExodusContext.uuidGenerator
 import utopia.flow.generic.ValueConversions._
 import utopia.nexus.http.Path
 import utopia.nexus.rest.Resource